v0.1.1.2 - light rebalance + FPS fix
 - now movement and other time-related effects not adjusted by FPS
 - hero sprite changed
 - fixed some orphographical mistakes
 - effect timeout changed from int to float
 - now start gold can be changed in config.cfg

+++ JustTest/Effect.h
@@ -13,7 +13,7 @@ enum EffectList {
 };
 
 struct Effect {
-	int time = 0;
+	float time = 0;
 	EffectList effect;
 
 	Effect(int time, EffectList effect) : time(time), effect(effect) {}
